Golden/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
python-dotenv==1.0.1
openpyxl==3.1.2
pytest==8.3.5
